vector.txt reading, tooltip fixes/improvements, others

+++ src/ini.h
@@ -7,7 +7,7 @@ extern struct Settings
 {
 	//default
 	float color[3]{0.324f, 0.409f, 0.185f};
-	int zoomLevel = 3;
+	float zoomLevel = 3.0f;
 	bool bilinear = 0;
 	int theme = 0;
 	float fontSize = 18.f;
